웹서비스/requirements.txt
pandas
numpy
torch
scikit-learn
matplotlib
tqdm
google-generativeai